pyproject.toml
[tool.poetry]
name = "kalshi-hft-bot"
version = "0.2.0"
description = "High-frequency trading bot for Kalshi prediction markets"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
websockets = "^12.0"
aiohttp = "^3.9.1"
cryptography = "^42.0.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
supabase = "^2.3.4"
structlog = "^24.1.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
